Added file via upload
plane location calculator with a variety of built in calculations. Done with the help of @ffeirou on my twitch stream. He helped with string formatting to allow me to do single parameter slow_type function on print outputs. Excepting his contribution and the slow_type function at the top(which I found on stackexchance as a forum response), all other work is of my creation.